+++ mesos_master/datadog_checks/mesos_master/mesos_master.py
@@ -160,6 +160,7 @@ class MesosMaster(AgentCheck):
             msg = str(e)
             status = AgentCheck.CRITICAL
         finally:
+            self.log.debug('Request to url : {0}, timeout: {1}, message: {2}'.format(url, timeout, msg))
             if self.service_check_needed:
                 self.service_check(self.SERVICE_CHECK_NAME, status, tags=tags,
                                    message=msg)
@@ -167,7 +168,7 @@ class MesosMaster(AgentCheck):
             if status is AgentCheck.CRITICAL:
                 self.service_check(self.SERVICE_CHECK_NAME, status, tags=tags,
                                    message=msg)
-                raise CheckException("Cannot connect to mesos, please check your configuration.")
+                raise CheckException('Cannot connect to mesos. Error: {0}'.format(msg))
 
         if r.encoding is None:
             r.encoding = 'UTF8'