+++ mesos_master/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Changes
 
+* [BUGFIX] Add debug logs and a noisier error message.
 * [FEATURE] adds custom tag support for service checks.
 
 1.1.0 / 2017-08-21